+++ ErsatzTV.Infrastructure/Data/Repositories/ChannelRepository.cs
@@ -14,44 +14,59 @@ namespace ErsatzTV.Infrastructure.Data.Repositories
     public class ChannelRepository : IChannelRepository
     {
         private readonly IDbConnection _dbConnection;
-        private readonly TvContext _dbContext;
+        private readonly IDbContextFactory<TvContext> _dbContextFactory;
 
-        public ChannelRepository(TvContext dbContext, IDbConnection dbConnection)
+        public ChannelRepository(IDbContextFactory<TvContext> dbContextFactory, IDbConnection dbConnection)
         {
-            _dbContext = dbContext;
+            _dbContextFactory = dbContextFactory;
             _dbConnection = dbConnection;
         }
 
         public async Task<Channel> Add(Channel channel)
         {
-            await _dbContext.Channels.AddAsync(channel);
-            await _dbContext.SaveChangesAsync();
+            await using TvContext dbContext = _dbContextFactory.CreateDbContext();
+            await dbContext.Channels.AddAsync(channel);
+            await dbContext.SaveChangesAsync();
             return channel;
         }
 
-        public Task<Option<Channel>> Get(int id) =>
-            _dbContext.Channels
+        public async Task<Option<Channel>> Get(int id)
+        {
+            await using TvContext dbContext = _dbContextFactory.CreateDbContext();
+            return await dbContext.Channels
                 .Include(c => c.Artwork)
+                .Include(c => c.Watermark)
                 .OrderBy(c => c.Id)
                 .SingleOrDefaultAsync(c => c.Id == id)
                 .Map(Optional);
+        }
 
-        public Task<Option<Channel>> GetByNumber(string number) =>
-            _dbContext.Channels
+        public async Task<Option<Channel>> GetByNumber(string number)
+        {
+            await using TvContext dbContext = _dbContextFactory.CreateDbContext();
+            return await dbContext.Channels
                 .Include(c => c.FFmpegProfile)
                 .ThenInclude(p => p.Resolution)
+                .Include(c => c.Artwork)
+                .Include(c => c.Watermark)
                 .OrderBy(c => c.Number)
                 .SingleOrDefaultAsync(c => c.Number == number)
                 .Map(Optional);
+        }
 
-        public Task<List<Channel>> GetAll() =>
-            _dbContext.Channels
+        public async Task<List<Channel>> GetAll()
+        {
+            await using TvContext dbContext = _dbContextFactory.CreateDbContext();
+            return await dbContext.Channels
                 .Include(c => c.FFmpegProfile)
                 .Include(c => c.Artwork)
                 .ToListAsync();
+        }
 
-        public Task<List<Channel>> GetAllForGuide() =>
-            _dbContext.Channels
+        public async Task<List<Channel>> GetAllForGuide()
+        {
+            await using TvContext dbContext = _dbContextFactory.CreateDbContext();
+            return await dbContext.Channels
                 .Include(c => c.Artwork)
                 .Include(c => c.Playouts)
                 .ThenInclude(p => p.Items)
@@ -80,23 +95,57 @@ namespace ErsatzTV.Infrastructure.Data.Repositories
                 .ThenInclude(i => (i as MusicVideo).Artist)
                 .ThenInclude(a => a.ArtistMetadata)
                 .ToListAsync();
+        }
 
-        public Task Update(Channel channel)
+        public async Task<bool> Update(Channel channel)
         {
-            _dbContext.Channels.Update(channel);
-            return _dbContext.SaveChangesAsync();
+            await using TvContext dbContext = _dbContextFactory.CreateDbContext();
+            dbContext.Entry(channel).State = EntityState.Modified;
+            if (channel.Watermark != null)
+            {
+                dbContext.Entry(channel.Watermark).State =
+                    channel.WatermarkId == null ? EntityState.Added : EntityState.Modified;
+            }
+
+            foreach (Artwork artwork in Optional(channel.Artwork).Flatten())
+            {
+                dbContext.Entry(artwork).State = artwork.Id > 0 ? EntityState.Modified : EntityState.Added;
+            }
+
+            bool result = await dbContext.SaveChangesAsync() > 0;
+            return result;
         }
 
         public async Task Delete(int channelId)
         {
-            Channel channel = await _dbContext.Channels.FindAsync(channelId);
-            _dbContext.Channels.Remove(channel);
-            await _dbContext.SaveChangesAsync();
+            await using TvContext dbContext = _dbContextFactory.CreateDbContext();
+            Channel channel = await dbContext.Channels.FindAsync(channelId);
+            dbContext.Channels.Remove(channel);
+            await dbContext.SaveChangesAsync();
         }
 
         public Task<int> CountPlayouts(int channelId) =>
             _dbConnection.QuerySingleAsync<int>(
                 @"SELECT COUNT(*) FROM Playout WHERE ChannelId = @ChannelId",
                 new { ChannelId = channelId });
+
+        public async Task<Unit> RemoveWatermark(Channel channel)
+        {
+            if (channel.Watermark != null)
+            {
+                await _dbConnection.ExecuteAsync(
+                    "UPDATE Channel SET WatermarkId = NULL WHERE Id = @ChannelId",
+                    new { ChannelId = channel.Id });
+
+                await _dbConnection.ExecuteAsync(
+                    "DELETE FROM ChannelWatermark WHERE Id = @WatermarkId",
+                    new { channel.WatermarkId });
+
+                channel.Watermark = null;
+                channel.WatermarkId = null;
+            }
+
+            return Unit.Default;
+        }
     }
 }