+++ ErsatzTV.Infrastructure/Migrations/20210610122252_Add_ChannelWatermark.Designer.cs
@@ -3,14 +3,16 @@ using System;
 using ErsatzTV.Infrastructure.Data;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
 namespace ErsatzTV.Infrastructure.Migrations
 {
     [DbContext(typeof(TvContext))]
-    partial class TvContextModelSnapshot : ModelSnapshot
+    [Migration("20210610122252_Add_ChannelWatermark")]
+    partial class Add_ChannelWatermark
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
@@ -446,6 +448,9 @@ namespace ErsatzTV.Infrastructure.Migrations
 
                     b.Property<int>("AudioSampleRate")
                         .HasColumnType("INTEGER");
+
+                    b.Property<string>("FrameRate")
+                        .HasColumnType("TEXT");
 
                     b.Property<int>("HardwareAcceleration")
                         .HasColumnType("INTEGER");